requirements.txt
pydub
jinja2
python-mpd2